Parse a string buffer
The spec does not specify the encoding for the pieces string, just that
it contains SHA1 hashes of length 20. When decoding a sample torrent
file as UTF-8 (the default encoding for buf.toString()), I ended up
with a pieces string shorter than the specified length.

SHA1 just returns a 160-bit hash. It is not guaranteed to only contain
valid UTF-8 characters.

> SHA1 [...] takes an input and produces a 160-bit (20-byte) hash value
> known as a message digest – typically rendered as 40 hexadecimal
> digits.

https://en.wikipedia.org/wiki/SHA-1

When we try to parse a hash to a string, and we encounter a byte
sequence that is not a valid UTF-8 character, those bytes will be
substituted with a single replacement character. Hence a shorter string.
For example, the first three bytes of the hex sequence f0 92 80 79 will
be replaced with the replacement character, because these four bytes
together are not a valid UTF-8 character. This chart is useful for
understanding the range of valid characters and how it's possible for an
invalid multi-byte sequence to exist:

https://design215.com/toolbox/utf8-4byte-characters.php

The spec does specify UTF-8 for other strings, so we can't specify a
single-byte encoding (i.e. ASCII) when decoding the torrent file buffer.
I don't think that would work anyway, because not all bytes would map to
characters. Working with the key values as buffers is the safe option.

buf[index] returns the decimal representation of the octet at the index,
which we can compare with a character code. charCodeAt() returns the
character code (UTF-16 code point) at the index, which we can use for
comparison. The UTF-16 code unit matches the Unicode code point for
code points which can be represented in a single UTF-16 code unit. The
first 128 Unicode code points are a direct match of the ASCII character
encoding.

https://nodejs.org/api/buffer.html#buftostringencoding-start-end
https://nodejs.org/api/buffer.html#bufindex
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/charCodeAt