Encode info dict
The tracker GET request must include the SHA1 hash of the info dict from
the torrent file. The Ubuntu tracker index provides info hashes for
verification: https://torrent.ubuntu.com/tracker_index.

There are at least a couple of ways we could approach this. For now,
take the decoded info dict, re-bencode it, and use that to generate the
hash. String object keys, which we have here, are ordered by insertion,
so we can depend on the hash of our re-encoded data being the same as
that of the data in the file.

I ran into an interesting bug caused by concatenating a buffer with
a string. When we converted the buffer to a string on concatenation,
invalid byte sequences were replaced with the three-byte replacement
character (ef bf bd). This added an extra two bytes for many characters,
increasing the length of the string.

Add a temporary entrypoint that prints the decoded data and the info
hash. Adjust the type definitions to allow passing a dict to the
encodeInfo function.

The next challenge is to URL-encode the hash value, which is required
before sending it to the tracker. The algorithm for that can be found in
this gist: https://gist.github.com/nkcmr/6917870. Basically, if a byte
is not an unreserved ASCII character, it needs to be converted to hex
and %-prefixed. Unreserved characters can be sent without encoding.