Add function to percent encode buffer
Algorithm taken from: https://gist.github.com/nkcmr/6917870.

For each byte in the input (a SHA1 hash of a bencoded info dictionary),
check if it is a character in the URI unreserved set. If so, add that
character to our string. If not, take the hex value, zero pad it if
required, and prefix with a percent symbol.