Make tracker get request
Spike out the tracker get request. We'll need to refactor this, of
course.

URLSearchParams are automatically percent-encoded. We don't want this
for our info hash, which is already encoded, so append that to the URL
manually.

https.get passes a readable stream to its callback, which emits a
'data' event when data is available to be read. More generally, Node
coordinates asynchronous operations via an event-driven architecture.
All objects that emit events are instances of EventEmitter, and expose
an on() function that allows us to attach listener functions to named
events.