Perform peer handshake
Very rough code, but it performs the handshake successfully.

\x is an escape sequence for hex codes. It allows us to include ASCII
characters in a string using their hex code.

The spec has:

> The handshake starts with character ninteen [sic] (decimal)

Here, decimal refers to the decimal value of the byte that represents
the ASCII character (001 0011). In hex, this is 13. The character itself
is a control code - presumably an arbitrary choice.

I think the next step should probably be to continue spiking out how to
send messages, before starting to think about the code structures we
need to implement this properly.