Fix compilation errors
Introduce some basic types and assert that our torrent file and tracker
data conform to these. This doesn't provide any runtime checking, so
we'll need to replace it with proper narrowing, perhaps in class
constructors. We have two options here: narrowing via control flow, or
via type predicates. [1]

For the Download type, we must first assert the type of the tracker
response data as unknown. I think this is because BDict does not specify
a property peers with type Peer[], and so the compiler is not confident
that Download is a more specific version of BDict (it works if we type
peers as one of Buffer, number, BList or BDict. [2]

unknown is a type-safe counterpart for any. Unlike any, we cannot assign
unknown to another type without an assertion or narrowing. [3]

[1]: https://www.typescriptlang.org/docs/handbook/2/narrowing.html
[2]: https://www.typescriptlang.org/docs/handbook/2/everyday-types.html#type-assertions
[3]: https://stackoverflow.com/a/51439876