Use type predicates for narrowing
https://www.typescriptlang.org/docs/handbook/2/narrowing.html#using-type-predicates

A function with a type predicate return type enables us to narrow the
type of its argument based on its return value.

I assumed that if I switched the type predicate return type for a
boolean then the type checker would still be satisfied, given the logic
is the same and the tests pass. This was incorrect.

This led to two discoveries: the value of the constructor property
does not narrow its owner's type, presumably because it is mutable.
instanceof does, but this is not as specific (arrays and other
prototypes are instances of Object). Also, type narrowing only applies
in the local scope, so you cannot extract logic to functions. I can't
find any documentation for this beyond Stack Overflow answers right now,
but it's definitely the behaviour I've observed.