Set bitfield and send interested message
TODO: We can't use async/await with the socket methods, because they do
not return Promises. Where we need to perform some action in response to
some event, e.g. send a handshake message on connect, we just listen for
those events. We could also pass a callback (listener) to connect().