Move piece request to Peer
This completes the move of the experimental code we'd written so far out
of the main function. The next thing to do will be to tighten up our
message processing logic (presumably we need to handle both incomplete
and concatenated messages), before we start thinking more about how to
request and store a complete piece.