Receive a message in multiple chunks
In the test we mock the socket write method, because the socket is not
open and writing to it fails with an error. Previously we've used spies
to do this, because we have needed to assert that the method has been
called.

We don't need to do that here. I tried to just overwrite the method with
a mock function, as below, but that failed with a type error. I'm not
totally clear on why that fails and spyOn doesn't, but I think it may be
because the latter wraps the function rather than replacing it, even if
you provide a mockImplementation.

  mockSocket.write = jest.fn<typeof mockSocket.write>();

For now, I'm just injecting a mock object that extends Socket and
overrides the write method. Perhaps that's better, but it would be
nice to understand why replacing the method with a mock fails, and to
standardise on a single approach.