Permit any reserved bytes in handshake message
Previously, per the spec, we only permitted zero in the reserved bytes
in the handshake message. We noticed that real peers would send bytes
other than zero, and so we would not accept their handshake messages. We
were also incorrectly sending the peer's own ID in our handshake (we now
send our own peer ID, referred to as client ID in the code).

Allow any reserved bytes in the handshake message, and send our peer ID.