Expose piece length
We will need this to dynamically calculate the number of 'chunks' to
fetch for a given piece.

We need to check the type of piece length because the type we pass to
construct an Info object (BDict) only specifies string keys and the
allowed data types. This is consistent with the approach we took when
returning peers for a Download, but we might want to consider how to
manage validation of torrent file and tracker response data.