WIP: Download a complete piece
This logic will need to be pulled out of Peer, but we might want to
spike a little more before deciding how to do that.

Concepts in play here:

* Peers list (each with a list of available pieces - bitfield)
* Pieces list (each with a list of chunks)

We'll want to be able to resume a download and assemble a pieces list
from objects on disk. It might be reasonable to keep chunks in memory
initially until we've assembled a complete piece, but I think ultimately
we'd want the same for these too.