Download from multiple peers
Set a max downloaders value, low for now to simplify manual testing. I
think this makes sense both because without it we could theoretically
exhaust the maximum number of available TCP connections, and because at
least one source recommends capping the number for performance reasons.
[1]

For now, we're both limiting the Peer interface available to the
Download class, and assuming that we will want a peer to proceed to
download autonomously. We may want to refactor both later if we find we
need more control.

[1]: https://wiki.theory.org/BitTorrentSpecification