Inject peer class dependency
Refactor Download to receive an injected peer class dependency. I much
prefer this as a technique for mocking that class, as opposed to using
Jest to override the implementation in place: it just feels much simpler
and affords much more control. We can still leverage Jest mocks by
setting them as methods on the mock class before we inject it.