Send disconnected peer to the back of the queue
We connect to up to maxDownloaders on download start. Some of these
connections may fail, or we may find that the peer does not have a piece
that we need. In either case, we will want to try connecting to other
peers, but possibly retry failed peers later. My first thought here
is to check the number of currently downloading peers at an interval,
and attempt new downloads if that number is below maxDownloaders. If
we treat the list of peers as a simple queue, moving previously failed
peers to the back, we should ensure that we prioritise new peers, and
allow some time to elapse before we retry.