Refactor Download tests
Make it possible to pass a collection of Peers into a Download. This
makes testing easier than the arrangement I had previously (injecting a
mock Peer class), because each peer can have its own mocks and state.
This will make future tests easier to write too.

The disconnect test was only working by accident. Both mock Peers
were set up to fire a disconnect event on download, which should have
resulted in the queue ending up in the same order as it started (both
peers being sent to the back in order). Instead, the disconnect event
listeners were called synchronously, leading to two attempts to download
from the first peer (i.e. the first disconnect listener was called
before the second download). I learned that when an EventEmitter object
emits an event, all of the functions attached to that specific event are
called synchronously. [1] In reality we would not see this behaviour,
because disconnect events would not be emitted within the download call.

[1]: https://nodejs.org/api/events.html#asynchronous-vs-synchronous