Add more test TODOs
The main point here is to ensure that we consider the bitfield when 1)
deciding whether to send an interested message and 2) selecting a piece
to request. For both we'll need some way of addressing individual bits
in a Buffer. A wild Bitfield class appears.

The lengths in my mock bitfield messages were incorrect. I'm guessing
these were just a truncated example response from manual testing.