Move Bitfield to separate package
This is mostly an experiment to learn a bit about npm workspaces. I
don't think there's any real advantage to merging this, although it
might make sense as a stepping stone on the way to extracting Bitfield
into a published npm package (another possible learning exercise).

Workspaces make it easy to add a dependency on a package in the same
repository, enabling a monorepo structure. Previously this would have
required npm link, which I have used before. npm link symlinks a package
to the global node_modules directory and then into the dependent
project. With a workspace, the package is symlinked into the project's
top-level node_modules directory. npm init -w will do this, but so will
an npm install after configuring the workspace manually.

TypeScript requires some further configuration. I've added a project
reference to the new package, which means that we must build the project
with the --build option to also build dependencies. We need to compile
the dependency in-place for it to be available to calling code in the
top-level project via its package.json entrypoint, so this is quite
convenient. Otherwise, we'd need to exclude it from our top-level
tsconfig and build it separately. There are alternative approaches to
managing a TypeScript monorepo. [1]

A couple of other things to note:

1) rootDir does not consider references, so in my case I saw errors
because the package files were outside of the src directory. Adding
an exclude declaration for the package directory in the top-level
tsconfig fixed this. I was surprised by this, because tsc knows to build
the referenced project separately. I suppose it's just a consequence
of rootDir being decoupled from include/exclude. I'm not sure why I
originally set this option, because it has no effect. It defaults to the
longest common path of all input files, which in this case is the src
directory anyway. [2]

2) If you inherit a base tsconfig in a subdirectory, any outDir set in
that file is relative to the directory in which that base configuration
is defined. Hence I've defined an outDir in each project config file.

[1]: https://www.reddit.com/r/javascript/comments/13cpiy1/askjs_is_there_a_silver_bullet_for_consuming/

(As far as I can tell, for the language server to detect references
in dependent packages, TypeScript must either be made aware of the
dependency via a project reference, or via a TypeScript entrypoint in
package.json, which introduces other challenges (the application must be
able to bundle TypeScript node_modules)).

[2]: https://www.typescriptlang.org/tsconfig#rootDir