Write piece file asynchronously
Our test is currently flaky, I think because there's a race condition
between the file read (in the test) and write (in the application)
callbacks. I experimented with a couple of workarounds. Retrying until
the file comparison succeeds seemed like a possible option, but I don't
think this will actually guarantee that the write file callback will be
executed before the test assertions that follow the read. I also tried
to find a way of watching for the file write to finish, without success:
there doesn't seem to be a way of listening for a file close event,
which I assume is what we'd want. So I think that we'll need to consider
an alternative approach to testing this, perhaps mocking the file write,
or finding some way to return a file write promise from the application
that we can await.