Rewrite tests for fetching and writing pieces
This addresses the test flakiness noted a couple of commits back.

Previously, I was testing the receive message event handler by emitting
a message event. Once I made this handler asynchronous (to write a
complete piece to disk) this was no longer practical: it introduced
race conditions between the test and application code. Test the handler
directly instead, which allows us to await it in the test.

We might want to switch to this pattern in other tests. A drawback is
that it is less integrated: we do not test the behaviour when an event
is emitted. We could add additional unit tests that verify that the
correct event listeners are in place.