Concatenate downloaded pieces
When the final piece is downloaded, concatenate all pieces into a single
file. A stream is a good choice here, because we don't want to hold the
entire downloaded file in memory.

Follow the approach to testing async handlers set out in 3c88c77.

I think this brings us to a complete implementation of the core download
functionality, although we currently only support single-file torrents.